requirements.txt
python-telegram-bot==20.8
APScheduler==3.10.4
requests
google-generativeai
Pillow